+++ NEWS.md
@@ -5,6 +5,7 @@
 * Added 'excluded_tabs' to the config file.
 * Added a footer with MCView and metacells versions.
 * Some modifications to the 'Genes' tab layout. 
+* Added an option to color the 2D projection by an axis from the Gene/Gene plot. 
 
 # MCView 0.2.7